refactor(api): extract the PR-preview lifecycle out of the sync route (cleanup 4/5)
routes/sync.ts carried ~190 loc of PR-preview domain logic (create/update/
tear-down on PR open/close, and the merge-time graduate that folds preview
versions + comment threads into the canonical doc). That's collection/blob
orchestration the webhook drives, not routing — it just happened to live in the
route file. Move it to lib/pr-preview.ts as makePrPreview({ meta, blobs,
baseUrl, launch }) -> { upsertPrPreview, previewCommentBody, removePrPreview,
graduatePreview }; the webhook calls the returned helpers unchanged. sync.ts
drops 952 -> 760. No behavior change; 618 tests green (SQLite).
